==> 0_Install_Packages.ipynb <==
### Install packages

!pip install git+https://github.com/huggingface/transformers

!git clone https://github.com/huggingface/transformers!python transformers/utils/download_glue_data.py

!python transformers/utils/download_glue_data.py

!pip install torch

!pip install twint

!pip install nest_asyncio
