==> 3_Feature_Extraction_&_Modeling.ipynb <==
### Import libraries

import numpy as np
import pandas as pd
import pickle

import torch
import transformers

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression, RidgeClassifier
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import RandomForestClassifier, StackingClassifier
from xgboost import XGBClassifier
from sklearn.metrics import f1_score, accuracy_score

import warnings
warnings.filterwarnings('ignore')

### User settings

PATH = "C:/Users/giuma/Dropbox/Web and Social media Analytics Project/"

do_FE = False # flag for the Feature Extraction process (set False when .pkl file can be imported)

### Import Cleaned Dataset

df_train = pd.read_pickle("train_nlp_disaster.pkl")
df_test = pd.read_pickle("test_nlp_disaster.pkl")

df_train.head()

### Feature Extraction

Use of DistilBERT (lighter version of BERT) for approaching Transfer Learning

# Load pretrained model/tokenizer for DistilBERT:
tokenizer = transformers.DistilBertTokenizer.from_pretrained('distilbert-base-uncased')
model = transformers.DistilBertModel.from_pretrained('distilbert-base-uncased')

def prepare_for_classifier(text, column = None):
    """
        Args:
        
        text: may be either a DataFrame (containing multiple tweets) or a string (single tweet)
    """
    
    ### Tokenization ###
    if ((isinstance(text, pd.DataFrame)) | (isinstance(text, pd.Series)))\
        & (column is not None):
        
        tokenized = text[column].apply(lambda x: tokenizer.encode(x, add_special_tokens=True))
        
    elif isinstance(text, str):
        
        tokenized = tokenizer.encode(text, add_special_tokens=True)
        tokenized = pd.Series([tokenized])  
    else:
        print('NOT RECOGNIZED DATA TYPE')
    
    # Output of tokenization is a pd.Series, with len equal to the number of tweets.
    # Each element of the series is a list which contains the tokenized tweet.
    
    
    ### Padding ###
    max_len = tokenized.apply(lambda x: len(x)).max()
    padded = np.array([i + [0]*(max_len-len(i)) for i in tokenized.values])
    
    
    ### Masking ###
    attention_mask = np.where(padded != 0 , 1, 0)
    attention_mask.shape
    
    # This mask contains 1 where values of the padded tokenized series is different from 0 (i.e. where no padding was applied)
    # and 0 elsewhere.
    
    
    ### Features extraction ###
    input_ids = torch.LongTensor(padded) #.tensor(padded)
    attention_mask = torch.LongTensor(attention_mask) #.tensor(attention_mask)

    with torch.no_grad():
        last_hidden_states = model(input_ids, attention_mask = attention_mask) # very long computation!!!
    
    features = last_hidden_states[0][:,0,:].numpy()
    
    # [:,0,:] -> allows to take only the features devoted to a classification task
    
    return features

if do_FE:
    # Extract features for classification (very heavy computation!!!)
    features = prepare_for_classifier(df_train, 'text_cleaned')

    # Save to pickle
    with open('features.pkl', 'wb') as handle:
        pickle.dump(features, handle, protocol = 4)

else:
    # Import pickle with features
    with open('features.pkl', 'rb') as handle:
        features = pickle.load(handle)


## Classification models

The derived features are used in order to train a series of classification models, which are finally ensembled by a stacking classifier (Supervised Learning)

labels = df_train['target_relabeled']

X_train, X_test, y_train, y_test = train_test_split(features, labels, random_state=42)

print(f"Features Train: {X_train.shape}")
print(f"Target Train: {y_train.shape}\n")

print(f"Features Test: {X_test.shape}")
print(f"Target Test: {y_test.shape}")

### Random Forest Classifier

model_rf = RandomForestClassifier()
parameters = {
    'bootstrap': [True],
    'ccp_alpha': [0.01, 0.001],
    'criterion': ["entropy"],
    'max_depth': [5,6,7,8,9,10],
    'max_features': [70],
    'min_samples_leaf': [1],
    'min_samples_split': [2],
    'class_weight': ['balanced'],
    'n_estimators': [100]
}

gs = GridSearchCV(model_rf, parameters, cv=3, scoring = 'f1', verbose=10, n_jobs=-1, refit=True)

gs = gs.fit(X_train, y_train)

print("** GRIDSEARCH RESULTS **")
print(f"Best score: {gs.best_score_} using {gs.best_params_}")

best_model = gs.best_estimator_
y_pred = best_model.predict(X_test)

y_pred_train = best_model.predict(X_train)
y_pred_test = best_model.predict(X_test)

print(f"** F1 Score **")
print(f"Train: {f1_score(y_train, y_pred_train):0.2f}")
print(f"Test: {f1_score(y_test, y_pred_test):0.2f}\n")

print(f"** Accuracy **")
print(f"Baseline - all zeroes: {accuracy_score(y_test, np.zeros(y_test.shape[0])):0.2f}")
print(f"Train: {accuracy_score(y_train, y_pred_train):0.2f}")
print(f"Test: {accuracy_score(y_test, y_pred_test):0.2f}")

### XGB Classifier

xgb_train = X_train
xgb_val = X_test[:952]
xgb_test = X_test[952:]

xgb_y_train = y_train
xgb_y_validation = y_test[:952]
xgb_y_test = y_test[952:]

xgb_model = XGBClassifier(learning_rate=0.1,
                          n_estimators=140,
                          max_depth=5,
                          reg_alpha=0.1,
                          reg_lambda=0.3,
                          min_child_weight=3,
                          gamma=0.2,
                          subsample=0.6,
                          colsample_bytree=1.0,
                          objective='binary:logistic',
                          nthread=4,
                          scale_pos_weight=1,
                          seed=42)
xgb_model.fit(xgb_train, 
              xgb_y_train,
              eval_set=[(xgb_train, xgb_y_train), (xgb_val, xgb_y_validation)], 
              verbose=20, 
              early_stopping_rounds=200)

xgb_train_pred = xgb_model.predict(xgb_train)
xgb_val_pred = xgb_model.predict(xgb_val)
xgb_test_pred = xgb_model.predict(xgb_test)

print(f'>>> XGB Model:')
print(f'Train f1: {f1_score(xgb_y_train, xgb_train_pred)}')
print(f'Validation f1: {f1_score(xgb_y_validation, xgb_val_pred)}')
print(f'Test f1: {f1_score(xgb_y_test, xgb_test_pred)}\n')

### Ridge Classifier

ridge_clf = RidgeClassifier()
ridge_params = {'alpha': np.arange(7,15),
                'class_weight': [None],
                'copy_X': [True],
                'fit_intercept': [True],
                'max_iter': [None],
                'normalize': [False],
                'random_state': [42],
                'solver': ['auto'],
                'tol': [0.001]}

ridge_gs = GridSearchCV(ridge_clf, ridge_params, cv=10, scoring = 'f1', verbose=10, n_jobs=-1, refit=True)

ridge_gs = ridge_gs.fit(X_train, y_train)

print("** GRIDSEARCH RESULTS **")
print(f"Best score: {ridge_gs.best_score_} using {ridge_gs.best_params_}")

best_ridge = ridge_gs.best_estimator_

y_pred_train = best_ridge.predict(X_train)
y_pred_test = best_ridge.predict(X_test)

print(f"** F1 Score **")
print(f"Train: {f1_score(y_train, y_pred_train):0.3f}")
print(f"Test: {f1_score(y_test, y_pred_test):0.3f}\n")

print(f"** Accuracy **")
print(f"Baseline - all zeroes: {accuracy_score(y_test, np.zeros(y_test.shape[0])):0.3f}")
print(f"Train: {accuracy_score(y_train, y_pred_train):0.3f}")
print(f"Test: {accuracy_score(y_test, y_pred_test):0.3f}")

### Naive Bayes

gnb_clf = GaussianNB()
gnb_clf.fit(X_train, y_train)

y_pred_train = gnb_clf.predict(X_train)
y_pred_test = gnb_clf.predict(X_test)

print(f"** F1 Score **")
print(f"Train: {f1_score(y_train, y_pred_train):0.3f}")
print(f"Test: {f1_score(y_test, y_pred_test):0.3f}\n")

print(f"** Accuracy **")
print(f"Baseline - all zeroes: {accuracy_score(y_test, np.zeros(y_test.shape[0])):0.3f}")
print(f"Train: {accuracy_score(y_train, y_pred_train):0.3f}")
print(f"Test: {accuracy_score(y_test, y_pred_test):0.3f}")

### Support Vector Classifier

svc_clf = SVC()
svc_params = {'C': [0.001, 0.01, 0.1, 1],
              'class_weight': [None],
              'coef0': [0.0],
              'degree': [3],
              'gamma': ['scale'],
              'kernel': ['rbf', 'linear'],
              'max_iter': [-1],
              'probability': [False],
              'random_state': [42],
              'shrinking': [True],
              'tol': [0.001],
              'verbose': [False]}

svc_gs = GridSearchCV(svc_clf, svc_params, cv=3, scoring = 'f1', verbose=10, n_jobs=-1, refit=True)

svc_gs = svc_gs.fit(X_train, y_train)

print("** GRIDSEARCH RESULTS **")
print(f"Best score: {svc_gs.best_score_} using {svc_gs.best_params_}")

best_svc = svc_gs.best_estimator_

y_pred_train = best_svc.predict(X_train)
y_pred_test = best_svc.predict(X_test)

print(f"** F1 Score **")
print(f"Train: {f1_score(y_train, y_pred_train):0.3f}")
print(f"Test: {f1_score(y_test, y_pred_test):0.3f}\n")

print(f"** Accuracy **")
print(f"Baseline - all zeroes: {accuracy_score(y_test, np.zeros(y_test.shape[0])):0.3f}")
print(f"Train: {accuracy_score(y_train, y_pred_train):0.3f}")
print(f"Test: {accuracy_score(y_test, y_pred_test):0.3f}")

### Ensemble - Stacking

estimators = [
    ('ridge', best_ridge),
    ('gnb', gnb_clf),
    ('svc', best_svc)
]
stacking = StackingClassifier(estimators=estimators,
                              final_estimator=LogisticRegression())

stacking.fit(X_train, y_train)

y_pred_train = stacking.predict(X_train)
y_pred_test = stacking.predict(X_test)

print(f"** F1 Score **")
print(f"Train: {f1_score(y_train, y_pred_train):0.2f}")
print(f"Test: {f1_score(y_test, y_pred_test):0.2f}\n")

print(f"** Accuracy **")
print(f"Baseline - all zeroes: {accuracy_score(y_test, np.zeros(y_test.shape[0])):0.2f}")
print(f"Train: {accuracy_score(y_train, y_pred_train):0.2f}")
print(f"Test: {accuracy_score(y_test, y_pred_test):0.2f}")

### Saving the model

with open('stacking_model.pkl', 'wb') as handle:
    pickle.dump(stacking, handle, protocol = 4)
